Chore: extract Hl7Time module to its own file
